refactor: simplify oxlint integration coverage
Amp-Thread-ID: https://ampcode.com/threads/T-01a016bb-8c8c-749a-bacd-b3a94b6d028e
